pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio-throttle>=1.0.2",
    "diskcache>=5.6.3",
    "google-genai>=1.29.0",
    "gradio>=5.42.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
]
